grades: refactor: create PostGrade model
