+++ src/app/components/grade/grade-statement/statement.component.ts
@@ -1,9 +1,9 @@
 import {Component, EventEmitter, Input, Output, ViewChild} from '@angular/core';
 import {ModalDirective} from 'ngx-bootstrap';
+
 import {IAppModal} from '../../shared/modal.interface';
-import {Grade} from '../../../models/Grade';
 import {GradeService} from '../../../services/grade.service';
-import {EmptyGrade} from '../EmptyGrade';
+import {EmptyGrade} from '../models/EmptyGrade';
 
 @Component({
     selector: 'app-grades-statement',
