+++ src/app/components/grade/models/EmptyGrade.ts
@@ -1,0 +1,16 @@
+export class EmptyGrade {
+  id?: number;
+  points: number;
+  empty: boolean;
+  courseId: number;
+  studentDegreeId: number;
+  onTime: boolean;
+
+  constructor(points: number, empty: boolean, courseId: number, studentDegreeId: number, onTime: boolean) {
+    this.points = points;
+    this.empty = empty;
+    this.courseId = courseId;
+    this.studentDegreeId = studentDegreeId;
+    this.onTime = onTime;
+  }
+}
